tests/wpt/tests/css/css-gaps/parsing/column-rule-visibility-items-invalid.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Gaps: parsing column-rule-visibility-items with invalid values</title>
  <link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-gaps-1/">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
  <div id="target"></div>
  <script>
    test_invalid_value('column-rule-visibility-items', 'auto');
    test_invalid_value('column-rule-visibility-items', 'true');
    test_invalid_value('column-rule-visibility-items', '10px');
    test_invalid_value('column-rule-visibility-items', 'default');
  </script>
</body>
</html>
